Heal cleanup combine XYList and RGBList
